Switched to a more proper perspective projection matrix for better precision as suggested by Arseny in Niagara
Frustum Culling is added. Currently, it is inefficiently done on the fly but will compute and store frustum planes before it is an easy task. It works, gives performance boost and correct results when camera is grazing chunks which was wrong before.

Still, the main bottleneck is the number of groups being dispatched. They are around 1.5-3M in worst cases and will be more with bigger data. Even the limit is 4M anyways and drivers scheduling them etc. is a very big overhead.

Next, I will make task shaders process more than one block in a chunk, in the beginning I will make them process all of them. This will increase the workload but reduce the number of groups greatly. I can also move testing the blocks individually to task shader and eliminate them if needed without an additional dispatch that is currently commented out in comp shader.