Parallelize loading times.
Make CTHead 8 bit too. We will need data compression as now I get 60 and more stable FPS  when used small chunks such as 4 and 8. This is because we can cull 80% of the chunks this way and get super fast performance

But, this comes with a cost of number of chunks. The memory bottleneck is not the data itself but the chunkmetadata but still there is no need for 32 bits of data while 8 bits work well