Done! Temporal Reprojection really solved the problem.
Of course this is not a perfect solution but most flickering and horrible stuttering is gone.

While camera is moving there are some false negatives but I can omit them. A perfect solution would be doing a second pass by refreshing the depth buffer and going through the blocks that have been rejected and testing them and rendering them. Unless really necessary, I won't do it but in a proper game this would be needed for example